[task/installer] Ground work for the installer.
Some initial work for the installer so that we can start tracking required
database changes for comming features. For now this lives in a seperate
branch without UMIL as the installer shouldnt be used for installing the
dev version. This will be merged into "develop/" when we dont expect any
real schema changes before the first public release.